+++ server/skillhub-domain/src/test/java/com/iflytek/skillhub/domain/label/LabelDefinitionServiceTest.java
@@ -22,7 +22,8 @@ class LabelDefinitionServiceTest {
     private final LabelDefinitionService service = new LabelDefinitionService(
             labelDefinitionRepository,
             labelTranslationRepository,
-            labelPermissionChecker
+            labelPermissionChecker,
+            100
     );
 
     @Test
