+++ web/src/features/namespace/use-my-namespaces.ts
@@ -2,4 +2,4 @@
  * Preserves a feature-local import path for dashboard namespace screens while
  * the underlying query implementation still lives in the shared hook module.
  */
-export { useMyNamespaces } from '@/shared/hooks/use-skill-queries'
+export { useMyNamespaces } from '@/shared/hooks/use-namespace-queries'
