+++ web/src/pages/dashboard.tsx
@@ -2,7 +2,7 @@ import { Link } from '@tanstack/react-router'
 import { useTranslation } from 'react-i18next'
 import { useAuth } from '@/features/auth/use-auth'
 import type { SkillSummary } from '@/api/types'
-import { useMySkills } from '@/shared/hooks/use-skill-queries'
+import { useMySkills } from '@/shared/hooks/use-user-queries'
 import { canViewGovernanceCenter } from '@/shared/lib/governance-access'
 import { getHeadlineVersion } from '@/shared/lib/skill-lifecycle'
 import { TokenList } from '@/features/token/token-list'
