+++ web/src/pages/dashboard/namespace-reviews.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next'
 import { formatLocalDateTime } from '@/shared/lib/date-time'
 import { Card } from '@/shared/ui/card'
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/shared/ui/tabs'
-import { useNamespaceDetail } from '@/shared/hooks/use-skill-queries'
+import { useNamespaceDetail } from '@/shared/hooks/use-namespace-queries'
 import { useReviewList } from '@/features/review/use-review-list'
 import { DashboardPageHeader } from '@/shared/components/dashboard-page-header'
 import { NamespaceHeader } from '@/features/namespace/namespace-header'
