+++ web/src/pages/dashboard/publish.tsx
@@ -13,7 +13,8 @@ import {
 } from '@/shared/ui/select'
 import { Label } from '@/shared/ui/label'
 import { Card } from '@/shared/ui/card'
-import { useMyNamespaces, usePublishSkill } from '@/shared/hooks/use-skill-queries'
+import { usePublishSkill } from '@/shared/hooks/use-skill-queries'
+import { useMyNamespaces } from '@/shared/hooks/use-namespace-queries'
 import { DashboardPageHeader } from '@/shared/components/dashboard-page-header'
 import { toast } from '@/shared/lib/toast'
 import { ApiError } from '@/api/client'
