+++ web/src/pages/search.tsx
@@ -9,7 +9,9 @@ import { SkillCard } from '@/features/skill/skill-card'
 import { SkeletonList } from '@/shared/components/skeleton-loader'
 import { EmptyState } from '@/shared/components/empty-state'
 import { Pagination } from '@/shared/components/pagination'
-import { useMyStars, useSearchSkills, useVisibleLabels } from '@/shared/hooks/use-skill-queries'
+import { useSearchSkills } from '@/shared/hooks/use-skill-queries'
+import { useVisibleLabels } from '@/shared/hooks/use-label-queries'
+import { useMyStars } from '@/shared/hooks/use-user-queries'
 import { normalizeSearchQuery } from '@/shared/lib/search-query'
 import { Button } from '@/shared/ui/button'
 import { APP_SHELL_PAGE_CLASS_NAME } from '@/app/page-shell-style'
