+++ web/src/shared/hooks/use-skill-queries.ts
@@ -1,37 +1,10 @@
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query'
-import type { SkillSummary, SkillDetail, SkillVersion, SkillVersionDetail, SkillFile, SearchParams, PagedResponse, PublishResult, Namespace, NamespaceMember, ManagedNamespace, CreateNamespaceRequest, NamespaceCandidateUser, NamespaceRole, LabelItem, LabelDefinition } from '@/api/types'
-import { fetchJson, fetchText, getCsrfHeaders, labelApi, meApi, namespaceApi, promotionApi, skillLifecycleApi, WEB_API_PREFIX } from '@/api/client'
-import { appendNamespaceMember, replaceNamespaceMemberRole } from '@/shared/lib/namespace-member-cache'
-import i18n from '@/i18n/config'
-import { buildSkillSearchUrl, shouldEnableNamespaceMemberCandidates } from './skill-query-helpers'
+import type { SkillSummary, SkillDetail, SkillVersion, SkillVersionDetail, SkillFile, SearchParams, PagedResponse, PublishResult } from '@/api/types'
+import { fetchJson, fetchText, getCsrfHeaders, skillLifecycleApi, WEB_API_PREFIX } from '@/api/client'
+import { getSkillDetailQueryKey } from './query-keys'
+import { buildSkillSearchUrl } from './skill-query-helpers'
 
-/**
- * Shared TanStack Query hooks for skill, namespace, and related dashboard data.
- *
- * This file currently acts as a broad query gateway for several features, centralizing cache keys,
- * backend fetchers, and invalidation rules used throughout the app.
- */
 const PUBLISH_REQUEST_TIMEOUT_MS = 60_000
-
-export function getI18nCacheKey() {
-  return i18n.resolvedLanguage || i18n.language || 'en'
-}
-
-export function getSkillDetailQueryKey(namespace: string, slug: string) {
-  return ['skills', namespace, slug, getI18nCacheKey()] as const
-}
-
-export function getVisibleLabelsQueryKey() {
-  return ['labels', 'visible', getI18nCacheKey()] as const
-}
-
-export function getSkillLabelsQueryKey(namespace: string, slug: string) {
-  return ['labels', 'skill', namespace, slug, getI18nCacheKey()] as const
-}
-
-export function getAdminLabelDefinitionsQueryKey() {
-  return ['labels', 'admin', getI18nCacheKey()] as const
-}
 
 async function searchSkills(params: SearchParams): Promise<PagedResponse<SkillSummary>> {
   return fetchJson<PagedResponse<SkillSummary>>(buildSkillSearchUrl(params))
@@ -40,26 +13,6 @@ async function searchSkills(params: SearchParams): Promise<PagedResponse<SkillSu
 async function getSkillDetail(namespace: string, slug: string): Promise<SkillDetail> {
   const cleanNamespace = namespace.startsWith('@') ? namespace.slice(1) : namespace
   return fetchJson<SkillDetail>(`${WEB_API_PREFIX}/skills/${cleanNamespace}/${slug}`)
-}
-
-async function getVisibleLabels(): Promise<LabelItem[]> {
-  return labelApi.listVisible()
-}
-
-async function getSkillLabels(namespace: string, slug: string): Promise<LabelItem[]> {
-  return labelApi.listSkillLabels(namespace, slug)
-}
-
-async function getAdminLabelDefinitions(): Promise<LabelDefinition[]> {
-  return labelApi.listAdminDefinitions()
-}
-
-async function attachSkillLabel(params: { namespace: string; slug: string; labelSlug: string }): Promise<LabelItem> {
-  return labelApi.attachSkillLabel(params.namespace, params.slug, params.labelSlug)
-}
-
-async function detachSkillLabel(params: { namespace: string; slug: string; labelSlug: string }): Promise<void> {
-  return labelApi.detachSkillLabel(params.namespace, params.slug, params.labelSlug)
 }
 
 async function getSkillVersions(namespace: string, slug: string): Promise<SkillVersion[]> {
@@ -83,59 +36,6 @@ async function getSkillDocumentation(namespace: string, slug: string, version: s
   return fetchText(`${WEB_API_PREFIX}/skills/${cleanNamespace}/${slug}/versions/${version}/file?path=${encodeURIComponent(path)}`)
 }
 
-async function getMySkills(params: { page?: number; size?: number; filter?: string } = {}): Promise<PagedResponse<SkillSummary>> {
-  return meApi.getSkills(params)
-}
-
-async function getMyStars(): Promise<SkillSummary[]> {
-  return meApi.getStars()
-}
-
-async function getMyStarsPage(params: { page?: number; size?: number } = {}): Promise<PagedResponse<SkillSummary>> {
-  return meApi.getStarsPage(params)
-}
-
-async function getMyNamespaces(): Promise<ManagedNamespace[]> {
-  return namespaceApi.listMine()
-}
-
-async function createNamespace(request: CreateNamespaceRequest): Promise<Namespace> {
-  return namespaceApi.create(request)
-}
-
-async function getNamespaceDetail(slug: string): Promise<Namespace> {
-  return namespaceApi.getDetail(slug)
-}
-
-async function getNamespaceMembers(slug: string): Promise<NamespaceMember[]> {
-  return namespaceApi.listMembers(slug)
-}
-
-async function submitPromotion(params: { sourceSkillId: number; sourceVersionId: number }): Promise<void> {
-  const globalNamespace = await namespaceApi.getDetail('global')
-  await promotionApi.submit({
-    sourceSkillId: params.sourceSkillId,
-    sourceVersionId: params.sourceVersionId,
-    targetNamespaceId: globalNamespace.id,
-  })
-}
-
-async function searchNamespaceMemberCandidates(params: { slug: string; search: string }): Promise<NamespaceCandidateUser[]> {
-  return namespaceApi.searchMemberCandidates(params.slug, params.search)
-}
-
-async function addNamespaceMember(params: { slug: string; userId: string; role: NamespaceRole }): Promise<NamespaceMember> {
-  return namespaceApi.addMember(params.slug, { userId: params.userId, role: params.role })
-}
-
-async function updateNamespaceMemberRole(params: { slug: string; userId: string; role: NamespaceRole }): Promise<NamespaceMember> {
-  return namespaceApi.updateMemberRole(params.slug, params.userId, params.role)
-}
-
-async function removeNamespaceMember(params: { slug: string; userId: string }): Promise<void> {
-  return namespaceApi.removeMember(params.slug, params.userId)
-}
-
 async function publishSkill(params: { namespace: string; file: File; visibility: string }): Promise<PublishResult> {
   const cleanNamespace = params.namespace.startsWith('@') ? params.namespace.slice(1) : params.namespace
   const formData = new FormData()
@@ -150,8 +50,6 @@ async function publishSkill(params: { namespace: string; file: File; visibility:
   })
 }
 
-// Query hooks stay close to the low-level fetchers so cache keys and invalidation rules remain
-// consistent across pages and feature wrappers.
 export function useSearchSkills(params: SearchParams) {
   return useQuery({
     queryKey: ['skills', 'search', params],
@@ -165,30 +63,6 @@ export function useSkillDetail(namespace: string, slug: string) {
     queryKey: getSkillDetailQueryKey(namespace, slug),
     queryFn: () => getSkillDetail(namespace, slug),
     enabled: !!namespace && !!slug,
-  })
-}
-
-export function useVisibleLabels(enabled = true) {
-  return useQuery({
-    queryKey: getVisibleLabelsQueryKey(),
-    queryFn: getVisibleLabels,
-    enabled,
-  })
-}
-
-export function useSkillLabels(namespace: string, slug: string, enabled = true) {
-  return useQuery({
-    queryKey: getSkillLabelsQueryKey(namespace, slug),
-    queryFn: () => getSkillLabels(namespace, slug),
-    enabled: enabled && !!namespace && !!slug,
-  })
-}
-
-export function useAdminLabelDefinitions(enabled = true) {
-  return useQuery({
-    queryKey: getAdminLabelDefinitionsQueryKey(),
-    queryFn: getAdminLabelDefinitions,
-    enabled,
   })
 }
 
@@ -224,93 +98,6 @@ export function useSkillVersionDetail(namespace: string, slug: string, version?:
   })
 }
 
-export function useMySkills(params: { page?: number; size?: number; filter?: string } = {}) {
-  return useQuery({
-    queryKey: ['skills', 'my', params],
-    queryFn: () => getMySkills(params),
-  })
-}
-
-export function useMyStars(enabled = true) {
-  return useQuery({
-    queryKey: ['skills', 'stars'],
-    queryFn: getMyStars,
-    enabled,
-  })
-}
-
-export function useMyStarsPage(params: { page?: number; size?: number } = {}, enabled = true) {
-  return useQuery({
-    queryKey: ['skills', 'stars', 'page', params],
-    queryFn: () => getMyStarsPage(params),
-    enabled,
-  })
-}
-
-export function useMyNamespaces() {
-  return useQuery({
-    queryKey: ['namespaces', 'my'],
-    queryFn: getMyNamespaces,
-  })
-}
-
-export function useCreateNamespace() {
-  const queryClient = useQueryClient()
-
-  return useMutation({
-    mutationFn: createNamespace,
-    onSuccess: (namespace) => {
-      queryClient.invalidateQueries({ queryKey: ['namespaces', 'my'] })
-      queryClient.invalidateQueries({ queryKey: ['namespaces', namespace.slug] })
-      queryClient.invalidateQueries({ queryKey: ['namespaces'] })
-    },
-  })
-}
-
-export function useNamespaceDetail(slug: string) {
-  return useQuery({
-    queryKey: ['namespaces', slug],
-    queryFn: () => getNamespaceDetail(slug),
-    enabled: !!slug,
-  })
-}
-
-export function useNamespaceMembers(slug: string) {
-  return useQuery({
-    queryKey: ['namespaces', slug, 'members'],
-    queryFn: () => getNamespaceMembers(slug),
-    enabled: !!slug,
-  })
-}
-
-export function useNamespaceMemberCandidates(slug: string, search: string, enabled = true) {
-  return useQuery({
-    queryKey: ['namespaces', slug, 'member-candidates', search],
-    queryFn: () => searchNamespaceMemberCandidates({ slug, search }),
-    enabled: shouldEnableNamespaceMemberCandidates(slug, search, enabled),
-  })
-}
-
-export function useSubmitPromotion() {
-  const queryClient = useQueryClient()
-
-  return useMutation({
-    mutationFn: submitPromotion,
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['promotions'] })
-      queryClient.invalidateQueries({ queryKey: ['governance'] })
-      queryClient.invalidateQueries({ queryKey: ['skills'] })
-    },
-  })
-}
-
-function invalidateNamespaceQueries(queryClient: ReturnType<typeof useQueryClient>, slug: string) {
-  queryClient.invalidateQueries({ queryKey: ['namespaces', 'my'] })
-  queryClient.invalidateQueries({ queryKey: ['namespaces', slug] })
-  queryClient.invalidateQueries({ queryKey: ['namespaces', slug, 'members'] })
-  queryClient.invalidateQueries({ queryKey: ['reviews'] })
-}
-
 export function usePublishSkill() {
   const queryClient = useQueryClient()
 
@@ -321,47 +108,6 @@ export function usePublishSkill() {
     },
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ['skills'] })
-    },
-  })
-}
-
-export function useAddNamespaceMember() {
-  const queryClient = useQueryClient()
-
-  return useMutation({
-    mutationFn: addNamespaceMember,
-    onSuccess: (member, variables) => {
-      queryClient.setQueryData<NamespaceMember[]>(
-        ['namespaces', variables.slug, 'members'],
-        (currentMembers) => appendNamespaceMember(currentMembers, member),
-      )
-      invalidateNamespaceQueries(queryClient, variables.slug)
-    },
-  })
-}
-
-export function useUpdateNamespaceMemberRole() {
-  const queryClient = useQueryClient()
-
-  return useMutation({
-    mutationFn: updateNamespaceMemberRole,
-    onSuccess: (member, variables) => {
-      queryClient.setQueryData<NamespaceMember[]>(
-        ['namespaces', variables.slug, 'members'],
-        (currentMembers) => replaceNamespaceMemberRole(currentMembers, variables.userId, member.role),
-      )
-      invalidateNamespaceQueries(queryClient, variables.slug)
-    },
-  })
-}
-
-export function useRemoveNamespaceMember() {
-  const queryClient = useQueryClient()
-
-  return useMutation({
-    mutationFn: removeNamespaceMember,
-    onSuccess: (_data, variables) => {
-      invalidateNamespaceQueries(queryClient, variables.slug)
     },
   })
 }
@@ -377,34 +123,6 @@ export function useArchiveSkill() {
       queryClient.invalidateQueries({ queryKey: ['skills', variables.namespace, variables.slug] })
       queryClient.invalidateQueries({ queryKey: ['skills', variables.namespace, variables.slug, 'versions'] })
       queryClient.invalidateQueries({ queryKey: ['skills'] })
-    },
-  })
-}
-
-function invalidateSkillLabelQueries(queryClient: ReturnType<typeof useQueryClient>, namespace: string, slug: string) {
-  queryClient.invalidateQueries({ queryKey: ['skills', namespace, slug] })
-  queryClient.invalidateQueries({ queryKey: ['labels', 'skill', namespace, slug] })
-  queryClient.invalidateQueries({ queryKey: ['skills'] })
-}
-
-export function useAttachSkillLabel() {
-  const queryClient = useQueryClient()
-
-  return useMutation({
-    mutationFn: attachSkillLabel,
-    onSuccess: (_data, variables) => {
-      invalidateSkillLabelQueries(queryClient, variables.namespace, variables.slug)
-    },
-  })
-}
-
-export function useDetachSkillLabel() {
-  const queryClient = useQueryClient()
-
-  return useMutation({
-    mutationFn: detachSkillLabel,
-    onSuccess: (_data, variables) => {
-      invalidateSkillLabelQueries(queryClient, variables.namespace, variables.slug)
     },
   })
 }
@@ -480,50 +198,6 @@ export function useRereleaseSkillVersion() {
       queryClient.invalidateQueries({ queryKey: ['skills', variables.namespace, variables.slug] })
       queryClient.invalidateQueries({ queryKey: ['skills', variables.namespace, variables.slug, 'versions'] })
       queryClient.invalidateQueries({ queryKey: ['skills'] })
-    },
-  })
-}
-
-export function useFreezeNamespace() {
-  const queryClient = useQueryClient()
-
-  return useMutation({
-    mutationFn: ({ slug }: { slug: string }) => namespaceApi.freeze(slug),
-    onSuccess: (_data, variables) => {
-      invalidateNamespaceQueries(queryClient, variables.slug)
-    },
-  })
-}
-
-export function useUnfreezeNamespace() {
-  const queryClient = useQueryClient()
-
-  return useMutation({
-    mutationFn: ({ slug }: { slug: string }) => namespaceApi.unfreeze(slug),
-    onSuccess: (_data, variables) => {
-      invalidateNamespaceQueries(queryClient, variables.slug)
-    },
-  })
-}
-
-export function useArchiveNamespace() {
-  const queryClient = useQueryClient()
-
-  return useMutation({
-    mutationFn: ({ slug, reason }: { slug: string; reason?: string }) => namespaceApi.archive(slug, reason),
-    onSuccess: (_data, variables) => {
-      invalidateNamespaceQueries(queryClient, variables.slug)
-    },
-  })
-}
-
-export function useRestoreNamespace() {
-  const queryClient = useQueryClient()
-
-  return useMutation({
-    mutationFn: ({ slug }: { slug: string }) => namespaceApi.restore(slug),
-    onSuccess: (_data, variables) => {
-      invalidateNamespaceQueries(queryClient, variables.slug)
     },
   })
 }
