+++ web/src/types/json.d.ts
@@ -1,4 +1,4 @@
 declare module '*.json' {
-  const value: any
+  const value: unknown
   export default value
 }
